Renames transform files
